refactor(pk): narrow on-disk PK to schema width, bump shard V6 + WAL V4 (Phase 2+3)
Remove Phase 1 shims. U64 PK tables now store 8B/row in shard files and WAL
blocks. U128 PK stays 16B/row. On-disk reads zero-extend to u128 via dynamic
pk_stride derived from schema.

- layout.rs: SHARD_VERSION 5→6
- shard_file.rs: drop padding adapters in build_shard_image/write_shard_streaming;
  XOR8 build uses stride = pk_sz/n with zero-extension to u128
- shard_reader.rs: add pk_stride field; fix get_pk, col_ptr_by_logical, slice_to_owned_batch
- gnitz-wire: WAL_FORMAT_VERSION 3→4
- wal_block.rs: add append_pk_region; encode/decode use schema pk_stride
- batch.rs: remove encode_to_wire, wire_byte_size, decode_from_wal_block shims
- runtime/wire.rs: schema_wal_block_size uses dynamic pk_stride (not hardcoded 16)

Cascade fix: shard_index.rs and wire.rs test helpers were passing 16B/row PKs
for U64 schemas; updated to 8B/row.

+++ crates/gnitz-engine/src/storage/shard_index.rs
@@ -696,7 +696,7 @@ mod tests {
         values: &[i64],
     ) -> String {
         let n = pks.len();
-        let pk_bytes: Vec<u8> = pks.iter().flat_map(|&p| (p as u128).to_le_bytes()).collect();
+        let pk_bytes: Vec<u8> = pks.iter().flat_map(|&p| p.to_le_bytes()).collect();
         let weights = vec![1i64; n];
         let nulls = vec![0u64; n];
         let blob: Vec<u8> = Vec::new();
